rename callback + import the new loader setup
