+++ alloc/alloc.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <flip_social.h>
-#include <callback/flip_social_callback.h>
+#include <callback/callback.h>
 #include <alloc/free.h>
 void auth_headers_alloc(void);
 FlipSocialFeedMini *flip_feed_info_alloc(void);
