+++ callback/callback.h
@@ -259,48 +259,12 @@ void flip_social_logged_in_explore_updated(void *context);
 void flip_social_logged_in_message_users_updated(void *context);
 
 // Add edits by Derek Jamison
-typedef enum DataState DataState;
-enum DataState
-{
-    DataStateInitial,
-    DataStateRequested,
-    DataStateReceived,
-    DataStateParsed,
-    DataStateParseError,
-    DataStateError,
-};
-
 typedef enum FlipSocialCustomEvent FlipSocialCustomEvent;
 enum FlipSocialCustomEvent
 {
     FlipSocialCustomEventProcess,
 };
 
-typedef struct DataLoaderModel DataLoaderModel;
-typedef bool (*DataLoaderFetch)(DataLoaderModel *model);
-typedef char *(*DataLoaderParser)(DataLoaderModel *model);
-struct DataLoaderModel
-{
-    char *title;
-    char *data_text;
-    DataState data_state;
-    DataLoaderFetch fetcher;
-    DataLoaderParser parser;
-    void *parser_context;
-    size_t request_index;
-    size_t request_count;
-    ViewNavigationCallback back_callback;
-    FuriTimer *timer;
-    FlipperHTTP *fhttp;
-};
-void flip_social_generic_switch_to_view(FlipSocialApp *app, char *title, DataLoaderFetch fetcher, DataLoaderParser parser, size_t request_count, ViewNavigationCallback back, uint32_t view_id);
-
-void flip_social_loader_draw_callback(Canvas *canvas, void *model);
-
-void flip_social_loader_init(View *view);
-
-void flip_social_loader_free_model(View *view);
-bool flip_social_custom_event_callback(void *context, uint32_t index);
 void messages_dialog_callback(DialogExResult result, void *context);
 void feed_dialog_callback(DialogExResult result, void *context);
 //
