+++ friends/flip_social_friends.h
@@ -1,7 +1,7 @@
 #ifndef FLIP_SOCIAL_FRIENDS_H
 #define FLIP_SOCIAL_FRIENDS_H
 #include "flip_social.h"
-#include <callback/flip_social_callback.h>
+#include <callback/callback.h>
 
 FlipSocialModel *flip_social_friends_alloc();
 bool flip_social_get_friends(FlipperHTTP *fhttp);
